fix: use float64 for Transaction.VolumeFactor
Avanza API returns 1.00 (a float) for volumeFactor, causing JSON
unmarshal to fail when the field was typed as *int.

+++ avanza/models/account.go
@@ -101,7 +101,7 @@ type Transaction struct {
 	ForeignTaxRate             *float64              `json:"foreignTaxRate"`
 	Isin                       *string               `json:"isin"`
 	Result                     *Value                `json:"result"`
-	VolumeFactor               *int                  `json:"volumeFactor"`
+	VolumeFactor               *float64              `json:"volumeFactor"`
 	Cancelled                  bool                  `json:"cancelled"`
 	CancelDate                 *string               `json:"cancelDate"`
 	VerificationNumber         string                `json:"verificationNumber"`
